requirements.txt
flask
Jinja2==3.0.3
parse
cryptography
python_dotenv
python-i18n
google-cloud-logging
google-cloud-storage
pymongo[srv]
flask-plugins
pylint==2.14.0
cerberus
pytest
mock
kafka-python
xlwt
pyjwt
requests
pendulum
pyotp
qrcode
httpagentparser
log4mongo
ipaddress
numpy
Flask-Mail
itsdangerous
flask-socketio
websocket-client
